logs and notification done
