+++ client/package.json
@@ -29,10 +29,12 @@
     "buffer": "^6.0.3",
     "crypto-browserify": "^3.12.0",
     "d3": "^7.9.0",
+    "esm": "^3.2.25",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.6.0",
     "https-browserify": "^1.0.0",
     "os-browserify": "^0.3.0",
+    "pkg-dir": "^4.2.0",
     "process": "^0.11.10",
     "react": "^18.3.1",
     "react-datepicker": "^7.2.0",
