+++ client/src/index.js
@@ -5,6 +5,7 @@ import { Provider } from 'react-redux';
 import { store, persistor } from '../src/redux/store.js';
 import { PersistGate } from 'redux-persist/integration/react';
 import { GoogleOAuthProvider } from '@react-oauth/google';
+import { WebSocketProvider } from '../src/redux/wsContext.js'; // ws global state
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
@@ -13,9 +14,11 @@ root.render(
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
         <GoogleOAuthProvider>
-          <App />
+          <WebSocketProvider>
+            <App />
+          </WebSocketProvider>
         </GoogleOAuthProvider>
-      </PersistGate> 
+      </PersistGate>
     </Provider>
   // </React.StrictMode>
 );
